added Smart Detector related code

+++ src/test/java/steps/Hooks.java
@@ -1,5 +1,6 @@
 package steps;
 
+import io.cucumber.java.Scenario;
 import org.junit.After;
 import org.junit.Before;
 import org.openqa.selenium.WebDriver;
@@ -8,21 +9,26 @@ import utilities.Driver;
 import java.util.concurrent.TimeUnit;
 
 public class Hooks {
-    WebDriver driver = Driver.getDriver();
+    WebDriver driver;
 
     @Before
-    public void setUp(){
-        driver.manage().window().maximize();
-        driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
-        driver.manage().deleteAllCookies();
-        System.out.println("Before Scenario Method");
+    public void setUp(Scenario scenario) {
+        if (!scenario.getSourceTagNames().contains("@api")) {
+            driver = Driver.getDriver();
+            driver.manage().window().maximize();
+            driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
+            driver.manage().deleteAllCookies();
+            System.out.println("Before Scenario Method");
+        }
+
     }
-
-
     @After
-    public void tearDown() throws InterruptedException {
-        Thread.sleep(1000);
-        driver.quit();
-        System.out.println("After Scenario Method");
+    public void tearDown(Scenario scenario) throws InterruptedException {
+        if (!scenario.getSourceTagNames().contains("@api")) {
+            Thread.sleep(1000);
+            driver.quit();
+            System.out.println("After Scenario Method");
+
+        }
     }
 }
